pyproject.toml
[tool.poetry]
name = "assignment3"
version = "0.1.0"
description = ""
authors = ["Christian Still <u19060123@tuks.co.za>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.28.68"
tqdm = "^4.66.1"
jupyter = "^1.0.0"
google = "^3.0.0"
lightning = "^2.1.0"
pandas = "1.4.3"
scikit-learn = "^1.3.2"
torchvision = "^0.16.0"
pillow = "^10.1.0"
ray = {extras = ["tune"], version = "^2.8.0"}


[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
